requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
python-dotenv
websockets
